Update barcode_faker.R
Hi, thank you for this great function! 

I had to add barcodes to ~1000 files and I got 10% of them with restriction sites so I added a checking step for that inside barcode_inventor(). I suggest using the restriction site as an argument with ApeKI as default. Maybe the replacement could also be an argument...

I also noted that barcode_inventor() does not work as a recursive function and in the case that barcodes are not unique it does not solve it. I could not work out how to fix it but anyway, giving an appropriate read_length will make barcodes unique...

I hope you find this addition useful!